+ 38.3 TT rewrite
one TT for both sides
TT cutoffs in only non-pv nodes
fixed edge case in hash key generation

+++ src/uci.rs
@@ -315,7 +315,7 @@ fn parse_perft(command: &str, position: &mut Board) {
     }
 }
 
-pub fn print_thinking(depth: usize, eval: i32, s: &Searcher, start: Instant) {
+pub fn print_thinking(depth: u8, eval: i32, s: &Searcher, start: Instant) {
     println!(
         "info depth {} score cp {} nodes {} pv{} time {} nps {}",
         depth,
